+++ saleor/checkout/migrations/0002_auto_20170206_0407.py
@@ -3,8 +3,6 @@
 from __future__ import unicode_literals
 
 import django.db.models.deletion
-import django_prices.models
-from django.conf import settings
 from django.contrib.postgres import fields
 from django.db import migrations, models
 
@@ -50,12 +48,8 @@ class Migration(migrations.Migration):
         migrations.AlterField(
             model_name="cart",
             name="total",
-            field=django_prices.models.MoneyField(
-                currency=settings.DEFAULT_CURRENCY,
-                decimal_places=2,
-                default=0,
-                max_digits=12,
-                verbose_name="total",
+            field=models.DecimalField(
+                decimal_places=2, default=0, max_digits=12, verbose_name="total"
             ),
         ),
         migrations.AlterField(