+++ saleor/graphql/discount/mutations.py
@@ -83,10 +83,11 @@ class BaseDiscountCatalogueMutation(BaseMutation):
 class VoucherInput(graphene.InputObjectType):
     type = VoucherTypeEnum(
         description=(
-            "Voucher type: PRODUCT, CATEGORY SHIPPING or ENTIRE_ORDER. "
-            "Deprecated fields: "
-            "PRODUCT, COLLECTION, CATEGORY use SPECIFIC_PRODUCT instead. "
-            "VALUE use ENTIRE_ORDER instead."
+            "Voucher type: PRODUCT, CATEGORY SHIPPING or ENTIRE_ORDER.\n"
+            "Deprecated fields:\n"
+            "\t1. PRODUCT, COLLECTION, CATEGORY use SPECIFIC_PRODUCT instead. "
+            "VALUE use ENTIRE_ORDER instead.\n"
+            "\t2. minAmountSpent use minSpentAmount instead"
         )
     )
     name = graphene.String(description="Voucher name.")
@@ -156,7 +157,12 @@ class VoucherCreate(ModelMutation):
         voucher_type = data.get("type", None)
         if voucher_type == VoucherTypeEnum.VALUE:
             data["type"] = VoucherTypeEnum.ENTIRE_ORDER.value
+
         cleaned_input = super().clean_input(info, instance, data)
+
+        min_spent_amount = cleaned_input.pop("min_amount_spent", None)
+        if min_spent_amount is not None:
+            cleaned_input["min_spent_amount"] = min_spent_amount
         return cleaned_input