+++ saleor/graphql/giftcard/mutations.py
@@ -52,8 +52,8 @@ class GiftCardCreate(ModelMutation):
         cleaned_input = super().clean_input(info, instance, data)
         balance = cleaned_input.get("balance", None)
         if balance:
-            cleaned_input["current_balance"] = balance
-            cleaned_input["initial_balance"] = balance
+            cleaned_input["current_balance_amount"] = balance
+            cleaned_input["initial_balance_amount"] = balance
         user_email = data.get("user_email", None)
         if user_email:
             try: