+++ saleor/graphql/product/mutations/products.py
@@ -439,8 +439,7 @@ class ProductInput(graphene.InputObjectType):
     )
     base_price = Decimal(description="Product price.")
     tax_rate = TaxRateType(
-        description="Tax rate.",
-        deprecation_reason=("taxRate is deprecated, Use taxCode"),
+        description="Tax rate.", deprecation_reason="taxRate is deprecated, Use taxCode"
     )
     tax_code = graphene.String(description="Tax rate for enabled tax gateway")
     seo = SeoInput(description="Search engine optimization fields.")
@@ -498,10 +497,10 @@ class ProductCreate(ModelMutation):
         # from the schema, only "basePrice" should be used here.
         price = data.get("base_price", data.get("price"))
         if price is not None:
-            cleaned_input["price"] = price
-            if instance.minimal_variant_price is None:
+            cleaned_input["price_amount"] = price
+            if instance.minimal_variant_price_amount is None:
                 # Set the default "minimal_variant_price" to the "price"
-                cleaned_input["minimal_variant_price"] = price
+                cleaned_input["minimal_variant_price_amount"] = price
 
         # FIXME  tax_rate logic should be dropped after we remove tax_rate from input
         tax_rate = cleaned_input.pop("tax_rate", "")
@@ -730,6 +729,14 @@ class ProductVariantCreate(ModelMutation):
     def clean_input(cls, info, instance, data):
         cleaned_input = super().clean_input(info, instance, data)
 
+        cost_price_amount = cleaned_input.pop("cost_price", None)
+        if cost_price_amount is not None:
+            cleaned_input["cost_price_amount"] = cost_price_amount
+
+        price_override_amount = cleaned_input.pop("price_override", None)
+        if price_override_amount is not None:
+            cleaned_input["price_override_amount"] = price_override_amount
+
         # Attributes are provided as list of `AttributeValueInput` objects.
         # We need to transform them into the format they're stored in the
         # `Product` model, which is HStore field that maps attribute's PK to
@@ -747,6 +754,7 @@ class ProductVariantCreate(ModelMutation):
                 raise ValidationError({"attributes": str(e)})
             else:
                 cleaned_input["attributes"] = attributes
+
         return cleaned_input
 
     @classmethod