+++ saleor/order/migrations/0009_auto_20151201_0820.py
@@ -1,8 +1,6 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
-import django_prices.models
-from django.conf import settings
 from django.db import migrations, models
 
 
@@ -17,9 +15,8 @@ class Migration(migrations.Migration):
         migrations.AddField(
             model_name="order",
             name="total_tax",
-            field=django_prices.models.MoneyField(
+            field=models.DecimalField(
                 decimal_places=2,
-                currency=settings.DEFAULT_CURRENCY,
                 max_digits=12,
                 blank=True,
                 null=True,