+++ saleor/shipping/migrations/0001_initial.py
@@ -3,8 +3,6 @@
 from __future__ import unicode_literals
 
 import django.db.models.manager
-import django_prices.models
-from django.conf import settings
 from django.db import migrations, models
 
 
@@ -304,11 +302,8 @@ class Migration(migrations.Migration):
                 ),
                 (
                     "price",
-                    django_prices.models.MoneyField(
-                        currency=settings.DEFAULT_CURRENCY,
-                        decimal_places=2,
-                        max_digits=12,
-                        verbose_name="price",
+                    models.DecimalField(
+                        decimal_places=2, max_digits=12, verbose_name="price"
                     ),
                 ),
                 (